# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Indo-German Mobility Corridor Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
...
  /candidates:
    get:
      operationId: listCandidates
      tags: [candidates]
      summary: List all candidates (admin/facilitator/employer/govt)
      parameters:
        - name: stage
          in: query
          schema:
            type: string
        - name: occupation
          in: query
          schema:
            type: string
        - name: district
          in: query
          schema:
            type: string
        - name: readinessMin
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Candidate list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Candidate"
  /candidates/me:
    get:
      operationId: getMyCandidateProfile
      tags: [candidates]
      summary: Get my candidate profile
      responses:
        "200":
          description: Candidate profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
    patch:
      operationId: updateMyCandidateProfile
      tags: [candidates]
      summary: Update my candidate profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CandidateUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"
...
  /candidates/{id}:
    get:
      operationId: getCandidate
      tags: [candidates]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Candidate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Candidate"

  # ── EMPLOYERS ────────────────────────────────────────────────────────────────
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: vacancyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"
    post:
      operationId: createApplication
      tags: [applications]
      summary: Apply to a vacancy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}:
    get:
      operationId: getApplication
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
    patch:
      operationId: updateApplication
      tags: [applications]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"

  # ── DOCUMENTS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
        role:
          type: string
    RegisterInput:
      type: object
      required: [email, password, name, role]
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [candidate, employer, trainer, facilitator, government, admin]
        phone:
          type: string
        organization:
          type: ["string", "null"]
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        phone:
          type: ["string", "null"]
        organization:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── CANDIDATE ──
    Candidate:
      type: object
      required: [id, userId, name, stage, readinessScore, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        district:
          type: ["string", "null"]
        state:
          type: ["string", "null"]
        occupation:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        experienceYears:
          type: ["integer", "null"]
        germanLanguageLevel:
          type: ["string", "null"]
        passportAvailable:
          type: boolean
        preferredGermanCity:
          type: ["string", "null"]
        stage:
          type: string
          enum: [discovery, registration, eligibility, language_training, qualification_recognition, document_vault, employer_matching, interview_offer, visa_readiness, post_arrival]
        readinessScore:
          type: integer
        profileComplete:
          type: boolean
        createdAt:
          type: string
    CandidateUpdate:
      type: object
      properties:
        district:
          type: string
        state:
          type: string
        occupation:
          type: string
        educationLevel:
          type: string
        experienceYears:
          type: integer
        germanLanguageLevel:
          type: string
        passportAvailable:
          type: boolean
        preferredGermanCity:
          type: string
...
    Application:
      type: object
      required: [id, candidateId, vacancyId, status, createdAt]
      properties:
        id:
          type: integer
        candidateId:
          type: integer
        vacancyId:
          type: integer
        candidateName:
          type: ["string", "null"]
        vacancyTitle:
          type: ["string", "null"]
        employerName:
          type: ["string", "null"]
        status:
          type: string
          enum: [applied, shortlisted, interview_scheduled, interview_done, offer_issued, offer_accepted, offer_rejected, withdrawn]
        appliedAt:
          type: string
        createdAt:
          type: string
    ApplicationInput:
      type: object
      required: [vacancyId]
      properties:
        vacancyId:
          type: integer
        coverNote:
          type: string
    ApplicationUpdate:
      type: object
      properties:
        status:
          type: string

    # ── DOCUMENT ──
...